move SiacoinOutputID to transaction.rs; remove serde_as hack, use type directly
