+++ src/encoding.rs
@@ -137,64 +137,6 @@ impl From<PublicKey> for PrefixedPublicKey {
     fn from(public_key: PublicKey) -> Self { PrefixedPublicKey(public_key) }
 }
 
-/// This wrapper allows us to use H256 internally but still serde as "scoid:" prefixed string
-#[derive(Clone, Debug, PartialEq)]
-pub struct SiacoinOutputId(pub Hash256);
-
-// FIXME this code pattern is reoccuring in many places and should be generalized with helpers or macros
-impl<'de> Deserialize<'de> for SiacoinOutputId {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct ScoidH256Visitor;
-
-        impl<'de> serde::de::Visitor<'de> for ScoidH256Visitor {
-            type Value = SiacoinOutputId;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                formatter.write_str("a string prefixed with 'scoid:' and followed by a 64-character hex string")
-            }
-
-            fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                if let Some(hex_str) = value.strip_prefix("scoid:") {
-                    Hash256::from_str(hex_str)
-                        .map(SiacoinOutputId)
-                        .map_err(|_| E::invalid_value(serde::de::Unexpected::Str(value), &self))
-                } else {
-                    Err(E::invalid_value(serde::de::Unexpected::Str(value), &self))
-                }
-            }
-        }
-
-        deserializer.deserialize_str(ScoidH256Visitor)
-    }
-}
-
-impl Serialize for SiacoinOutputId {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&format!("{}", self))
-    }
-}
-
-impl fmt::Display for SiacoinOutputId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result { write!(f, "h:{}", self.0) }
-}
-
-impl From<SiacoinOutputId> for Hash256 {
-    fn from(sia_hash: SiacoinOutputId) -> Self { sia_hash.0 }
-}
-
-impl From<Hash256> for SiacoinOutputId {
-    fn from(h256: Hash256) -> Self { SiacoinOutputId(h256) }
-}
-
 impl Encodable for Hash256 {
     fn encode(&self, encoder: &mut Encoder) { encoder.write_slice(&self.0); }
 }
